% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_deviance_lf.R
\name{delta_deviance_lf}
\alias{delta_deviance_lf}
\title{Calculate delta deviance of logistic fits of a null and alternative model}
\usage{
delta_deviance_lf(X, LF0, LF1)
}
\arguments{
\item{X}{A matrix of SNP genotypes, i.e. an integer matrix of 0's,
1's, 2's and \code{NA}s.
BEDMatrix is supported.
Sparse matrices of class Matrix are not supported (yet).}

\item{LF0}{Logistic factors for null model.}

\item{LF1}{Logistic factors for alternative model.}
}
\value{
The vector of delta deviance values, one per locus of \code{X}.
}
\description{
This function fits, at each locus of a given genotype matrix, two logistic models, and under the assumption that the models are nested, calculates the delta deviance between the two.
This general function is intended for testing models in a broad setting; for the specific problem of genetic association, the interface in \code{\link[=gcat]{gcat()}} and \code{\link[=gcat.stat]{gcat.stat()}} are more user-friendly.
}
\examples{
\dontrun{
devdiff <- delta_deviance_lf( X, LF0, LF1 )
}

}
