% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_geno}
\alias{sim_geno}
\title{Simulated data from PSD model}
\format{a matrix of 0's, 1's and 2's for the genotypes}
\usage{
sim_geno
}
\value{
simulated genotype matrix
}
\description{
10,000 SNPs, 1,000 individuals, first five SNPs are associated.
}

