% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_trait}
\alias{sim_trait}
\title{Simulated data from PSD model}
\format{a vector of traits}
\usage{
sim_trait
}
\value{
simulated traits
}
\description{
10,000 SNPs, 1,000 individuals, first five SNPs are associated.
}

